unreal-tek: Add support for PoE found on RTL838x switches
The common theme with PoE on RTL838x-based switches is a
microcontroller between the PSE chips and the CPU. Comms is done over
a UART, which has so far been consistently "ttyS1". The RTL838x chip
only has two UARTs, the first being used for serial console.

This profile focuses on controllers with 8 PoE ports. Devices with
fewer, or more ports exist in the wild, but supporting them is beyond
the scope of this change.

Some MCU firmware versions will export details about the PSE chips,
but this is hit or miss depending on the vendor. For example, on the
Engenius EWS2910P, the num_pse field is set to 0xff, indicating this
byte is not even used in the protocol. Thus PSE ID is not implemented.