poemgr: Sanity check profile before using it
A profile needs to meet certain conditions. It must have all its
member functions, and the number of chips and ports must fall within
a pre-allocated array. If any of those conditions are not met,
assumptions made elsewhere in poemgr cause hard-to debug crashes.

To resolve this temporarily, run a profile through a sanity check, and
make sure these conditions are met. Otherwise print a short message
directing the developer where to look next.