poemgr: Use floating point for power measurements
PoE standards specify power granularity to 0.1W. That can be clumsy to
represent with integers. Instead, use floating point and JSON doubles.

Note that json_object_new_double() will add a lot of pointless digits
to the output. This is resolved with json_object_new_double_s. The
format is kept at one digit after the decimal. This is sufficient for
power measurements.