poemgr: Treat profiles as const data structures
The way "profile" is used in poemgr implies to me that it is a
descriptor of some sorts. Intuitively, such descriptor would be const.
This is not the case for two reasons:
    1. The profile stores a "priv" pointer for context-specific data.
    2. The "pse_chips" array is stored in the profile.

Both "priv" and "pse_chips" contain dynamic data. It seems logical to
store dynamic data in the "poemgr_context" struct instead. To me, this
intuitive compared to storing these in the profile.

Therefore, move dynamic data out of the context, and into the profile,
and declare all profiles const.