poemgr: Do not exit poemgr_load_settings() using "goto"
gotos are evil. While the kernel likes to play with fire and use goto
for very specific and limited purposes, poemgr isn't the kernel.

In the case of poemgr_load_settings(), one variable and seven lines of
code are eliminated. The goto are not justified here. Replace them
with return statements.