poemgr: Make sure unconfigured ports are disabled
The ctx->ports array is initialized to '0'. Every port's settings will
start with "disabled == 0". When a port is not declared in the UCI
config, poemgr_load_port_settings() will not modify it. Thus, any
unconfigured port supported by the profile will remain enabled.

This is undesirable because every unconfigured entry is equivalent to
port 0 with all zero entries. This can likely overwrite the settings
for port 0.

To resolve this, disable every port before parsing the config. Then,
uncofigured ports will remain disabled, and should not affect the
device configuration which happens later.